Remove google-collections dependency
google-collections is old and has been succeeded by google-guava. It was
only used in unit tests cases for the convenience of constructing small
maps. These can be mostly replaced with Collections.singletonMap. Having
it as a compile dependency meant that this was propagated to dependent
projects.

Theoretically a dependent project could be relying on this transitive
dependency and break when it is removed, so I think this should be a
minor version bump.